Rename runtime to execution environment class
