+++ src/miv_simulator/simulator/__init__.py
@@ -1,6 +1,5 @@
 __doc__ = """Contains the end-user public API of the MiV-Simulator"""
 
-from miv_simulator.simulator import nh5
 from miv_simulator.simulator.distribute_synapses import distribute_synapses
 from miv_simulator.simulator.generate_connections import generate_connections
 from miv_simulator.simulator.generate_network_architecture import (
@@ -10,6 +9,6 @@ from miv_simulator.simulator.generate_synapse_forest import (
     generate_synapse_forest,
 )
 from miv_simulator.simulator.measure_distances import measure_distances
-from miv_simulator.simulator.runtime import Runtime
+from miv_simulator.simulator.execution_environment import ExecutionEnvironment
 from miv_simulator.utils.io import create_neural_h5
 from miv_simulator.utils.neuron import configure_hoc
